+++ netbox_contract/__init__.py
@@ -5,7 +5,7 @@ class ContractsConfig(PluginConfig):
     name = 'netbox_contract'
     verbose_name = 'Netbox contract'
     description = 'Contract management plugin for Netbox'
-    version = '2.4.1'
+    version = '2.4.4'
     author = 'Marc Lebreuil'
     author_email = 'marc@famillelebreuil.net'
     base_url = 'contracts'
@@ -18,6 +18,15 @@ class ContractsConfig(PluginConfig):
         'mandatory_invoice_fields': [],
         'hidden_invoice_fields': [],
         'mandatory_dimensions': [],
+        'supported_models': [
+            'circuits.circuit',
+            'circuits.virtualcircuit',
+            'dcim.site',
+            'dcim.device',
+            'dcim.rack',
+            'virtualization.virtualmachine',
+            'virtualization.cluster',
+        ],
     }