+++ CHANGELOG.md
@@ -5,6 +5,21 @@
 
 ## Version 2
 
+### Version 2.4.4
+* [288](https://github.com/mlebreuil/netbox-contract/issues/288) Add the possibility to assign contract any object type. By default the following objects types: 'circuits.circuit', 'circuits.virtualcircuit', 'dcim.site', 'dcim.device', 'dcim.rack', 'virtualization.virtualmachine', 'virtualization.cluster', 'ipam.ipaddress', 'ipam.prefix'. This list can be overriden within the PLUGINS_CONFIG configuration parameter. Check the README file or [documentation](https://mlebreuil.github.io/netbox-contract/) for more information.
+
+### Version 2.4.3
+
+* [282](https://github.com/mlebreuil/netbox-contract/issues/282) Add the possibility to assign contract to clusters and racks.
+* [276](https://github.com/mlebreuil/netbox-contract/issues/276) Add the possibility to update the provider in bulk for contracts.
+* [275](https://github.com/mlebreuil/netbox-contract/issues/275) Add the possibility to filter contract by provider an service provider. Remove the default value for the currency and internal party fields in the contract search form.
+* [274](https://github.com/mlebreuil/netbox-contract/issues/274) List contracts on providers and service providers pages
+
+### Version 2.4.2
+
+* [272](https://github.com/mlebreuil/netbox-contract/issues/272) Add missing contract_type fields to the api.
+
+
 ### Version 2.4.1
 
 * [257](https://github.com/mlebreuil/netbox-contract/issues/257) Add the possibility to filter invoices by accounting dimensions.